Finished SSL tests
